CC-55 Implement a JUnit Rule for test support
This makes it easier to use Constretto's support for handling @Tags
on a test class. Several JUnit Rules can be used at the same time,
while JUnit Runners (like ConstrettoJUnit4ClassRunner) is harder to
compose.

JUnit got support for Rules in version 4.7, so the JUnit dependency
has been upgraded to this.