Proof of concept for adding remote configuration stores to Constretto
A problem with the current ConfigurationStores is that it parses the whole configuration at once. When doing service discovery
it would be ideal to only perform a query when a key is needed.

This suggested change add a concept of schemas for remote stores, to segregate them from regular keys. This makes it easier
to only perform requests to the service discovery backend when necessary. This could of course bind the application more to
the specific service discovery technlogy, but by using interpolation it is possible to use regular keys as still query
the service discovery. By using interpolation it is also easier to prevent calls to a service discovery mechanism where
it is not available, eg. when developing locally.

Examples:

    config.evaluateToString("consul:service.backend.first.ip");

Will query Consul for an ip via a not yet implemented consul-store

    backend.url=http://#{consul:service.backend.first.ip}/
    @dev.backend.url=http://localhost:3000/

Will use service discovery as default, but will use a static address on local development

This is a proof of concept, so it is missing documentation.