Proposed feature: Property key from String field value
Currently, field name is used as key when @Configuration annotation has no value.

A benefit of this is that the key is clearly visible everywhere the field is accessed.

It cannot work, however, if the key consists of periods, like my.example.property=X

Which is why I propose support for using field value as key expression.

In most (?) IDEs, the key expression can then be viewed with Ctrl+mouseover or similar.

This also means less duplication of key expressions.

There are some disadvantages, like only working for String, which is why I won't be too sad if this feature request is not accepted :-)